+++ lib/stm32/f3/Makefile
@@ -43,7 +43,7 @@ OBJS            += gpio_common_all.o gpio_common_f0234.o \
 		   iwdg_common_all.o spi_common_all.o dma_common_l1f013.o\
 		   timer_common_all.o timer_common_f234.o flash_common_f234.o \
 		   flash.o exti_common_all.o rcc_common_all.o spi_common_f03.o
-OBJS		+= adc_common_v2.o
+OBJS		+= adc_common_v2.o adc_common_v2_multi.o
 
 OBJS		+= usb.o usb_control.o usb_standard.o
 OBJS		+= st_usbfs_core.o st_usbfs_v1.o
